backend/requirements-ingest.txt
# Requirements for local data ingestion script (ingest.py and test_ingest.py)
# This includes document loaders and the embedding model.

langchain-community
langchain-core
langchain-google-genai
qdrant-client
langchain-qdrant
python-dotenv
beautifulsoup4
lxml
google-generativeai
fastembed
